+++ cookbooks/chub-browserstack-tunnel/roles/browserstack-tunnel.rb
@@ -1,2 +1,0 @@
-name "browserstack-tunnel"
-run_list "recipe[chub-browserstack-tunnel::default]"