+++ cookbooks/chub_browserstack_tunnel/metadata.rb
@@ -1,0 +1,11 @@
+name             'chub_browserstack_tunnel'
+maintainer       'Kyle Titus'
+maintainer_email 'ktitus@commercehub.com'
+license          'All rights reserved'
+description      'Installs/Configures chub_browserstack_tunnel'
+long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
+version          '0.1.2'
+
+supports "ubuntu"
+depends "apt"
+depends "ark"