+++ roles/plaza.rb
@@ -1,3 +1,3 @@
 name "plaza"
 description "A node for the Plaza service"
-run_list "recipe[base]", "recipe[chub_sensu_sso::client]", "recipe[chub-plaza]"
+run_list "recipe[base]", "recipe[chub_sensu_sso::client]", "recipe[chub_plaza]"